+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## 0.2.0
 
-*2026-01-23*
+*2026-01-24*
 
 **Added**
 
@@ -14,6 +14,7 @@
 
 - Non-width image keywords starting with `$1` (e.g., `$1页` in Chinese)
 - Missing styles for the [signatureHelp](./README.md#signaturehelp) tooltip in the MediaWiki mode
+- Highlighting of parser functions and parser function argument names in the MediaWiki mode
 
 **Changed**